--- a/src/backend/catalog/pg_depend.c
+++ b/src/backend/catalog/pg_depend.c
@@ -28,6 +28,7 @@
 #include "catalog/pg_foreign_server.h"
 #include "catalog/pg_language.h"
 #include "catalog/pg_namespace.h"
+#include "catalog/pg_opclass.h"
 #include "catalog/pg_operator.h"
 #include "catalog/pg_proc.h"
 #include "catalog/pg_ts_dict.h"
@@ -872,6 +873,10 @@ depLockAndCheckObject(const ObjectAddress *referenced)
 		case OperatorRelationId:
 			cacheId = OPEROID;
 			objName = "operator";
+			break;
+		case OperatorClassRelationId:
+			cacheId = CLAOID;
+			objName = "operator class";
 			break;
 		default:
 	}