--- a/src/test/isolation2/expected/dependency.out
+++ b/src/test/isolation2/expected/dependency.out
@@ -1,5 +1,5 @@
--- Test different cases when a parallel transaction drops a dependency object
--- before current transaction is committed.
+-- Test cases when a parallel transaction drops a dependency object
+-- while current transaction is yet not committed.
 
 -- start_matchsubs
 -- m/cache lookup failed for type \d+/
@@ -18,7 +18,7 @@
 -- s/ruleutils.c:\d+/ruleutils.c:XXX/
 -- end_matchsubs
 
--- Case 1. Dependency on the schema.
+-- Case 1. Function dependency on the schema.
 create schema test_1_schema;
 CREATE
 
@@ -64,7 +64,7 @@ ERROR:  schema "test_1_schema" does not exist  (seg0 127.0.1.1:6002 pid=2715327)
 1: end;
 END
 
--- Case 2. Dependency on the return type.
+-- Case 2. Function dependency on the return type.
 create type test_2_type as (a int);
 CREATE
 
@@ -115,7 +115,7 @@ CONTEXT:  SQL function "test_2_function"
 1: end;
 END
 
--- Case 3. Dependency on the parameter type.
+-- Case 3. Function dependency on the parameter type.
 create type test_3_type as enum ('one', 'two');
 CREATE
 
@@ -162,7 +162,7 @@ CONTEXT:  SQL function "test_3_function"
 1: end;
 END
 
--- Case 4. Dependency on the language.
+-- Case 4. Function dependency on the language.
 create language plpythonu;
 CREATE
 
@@ -209,7 +209,7 @@ ERROR:  cache lookup failed for language XXXXX (fmgr.c:XXX)
 1: end;
 END
 
--- Case 5. Dependency on the parameter default expression.
+-- Case 5. Function dependency on the parameter default expression.
 create function test5_default_value_function() returns text as $$ select 'test'::text; /**/ $$ language sql;
 CREATE
 
@@ -257,7 +257,7 @@ HINT:  No function matches the given name and argument types. You might need to 
 1: end;
 END
 
--- Case 6. Dependency on the column default expression.
+-- Case 6. Table dependency on the column default expression.
 create function test_6_default_value_function() returns text as $$ select 'test'::text; /**/ $$ language sql;
 CREATE
 
@@ -308,7 +308,7 @@ ERROR:  cache lookup failed for function XXXXX (ruleutils.c:10062)  (seg2 127.0.
 1: end;
 END
 
--- Case 7. Dependency on the column type.
+-- Case 7. Table dependency on the column type.
 create type test_7_type as enum ('one', 'two');
 CREATE