--- a/src/test/isolation2/expected/dependency.out
+++ b/src/test/isolation2/expected/dependency.out
@@ -1,23 +1,6 @@
 -- Test cases when a parallel transaction drops a dependency object
 -- while current transaction is yet not committed.
 
--- start_matchsubs
--- m/cache lookup failed for type \d+/
--- s/cache lookup failed for type \d+/cache lookup failed for type XXXXX/
--- m/cache lookup failed for language \d+/
--- s/cache lookup failed for language \d+/cache lookup failed for language XXXXX/
--- m/cache lookup failed for function \d+/
--- s/cache lookup failed for function \d+/cache lookup failed for function XXXXX/
--- m/format_type.c:\d+/
--- s/format_type.c:\d+/format_type.c:XXX/
--- m/lsyscache.c:\d+/
--- s/lsyscache.c:\d+/lsyscache.c:XXX/
--- m/fmgr.c:\d+/
--- s/fmgr.c:\d+/fmgr.c:XXX/
--- m/ruleutils.c:\d+/
--- s/ruleutils.c:\d+/ruleutils.c:XXX/
--- end_matchsubs
-
 -- Case 1. Function dependency on the schema.
 create schema test_1_schema;
 CREATE
@@ -60,7 +43,7 @@ DROP
 2: commit;
 COMMIT
 1<:  <... completed>
-ERROR:  schema "test_1_schema" does not exist  (seg0 127.0.1.1:6002 pid=2715327)
+ERROR:  namespace ##### was concurrently dropped
 1: end;
 END
 
@@ -110,8 +93,7 @@ DROP
 2: commit;
 COMMIT
 1<:  <... completed>
-ERROR:  cache lookup failed for type XXXXX (format_type.c:XXX)
-CONTEXT:  SQL function "test_2_function"
+ERROR:  type ##### was concurrently dropped
 1: end;
 END
 
@@ -157,8 +139,7 @@ DROP
 2: commit;
 COMMIT
 1<:  <... completed>
-ERROR:  cache lookup failed for type XXXXX (lsyscache.c:XXX)
-CONTEXT:  SQL function "test_3_function"
+ERROR:  type ##### was concurrently dropped
 1: end;
 END
 
@@ -205,7 +186,7 @@ DROP
 2: commit;
 COMMIT
 1<:  <... completed>
-ERROR:  cache lookup failed for language XXXXX (fmgr.c:XXX)
+ERROR:  language ##### was concurrently dropped
 1: end;
 END
 
@@ -252,8 +233,7 @@ DROP
 2: commit;
 COMMIT
 1<:  <... completed>
-ERROR:  function test5_default_value_function() does not exist  (seg0 127.0.1.1:6002 pid=2718914)
-HINT:  No function matches the given name and argument types. You might need to add explicit type casts.
+ERROR:  procedure ##### was concurrently dropped
 1: end;
 END
 
@@ -304,7 +284,7 @@ DROP
 2: commit;
 COMMIT
 1<:  <... completed>
-ERROR:  cache lookup failed for function XXXXX (ruleutils.c:10062)  (seg2 127.0.1.1:6004 pid=2718915) (ruleutils.c:10062)
+ERROR:  procedure ##### was concurrently dropped
 1: end;
 END
 
@@ -352,7 +332,7 @@ DROP
 2: commit;
 COMMIT
 1<:  <... completed>
-ERROR:  type "test_7_type" does not exist  (seg0 127.0.1.1:6002 pid=2718914)
+ERROR:  type ##### was concurrently dropped
 1: end;
 END
 
@@ -403,6 +383,6 @@ DROP
 2: commit;
 COMMIT
 1<:  <... completed>
-ERROR:  collation "test_8_collation" for encoding "UTF8" does not exist  (seg0 127.0.1.1:6002 pid=2718914)
+ERROR:  collation ##### was concurrently dropped
 1: end;
 END