--- a/src/test/isolation2/expected/dependency.out
+++ b/src/test/isolation2/expected/dependency.out
@@ -386,3 +386,50 @@ COMMIT
 ERROR:  collation ##### was concurrently dropped
 1: end;
 END
+
+-- Case 9. Text search configuration dependency on the parser.
+create text search parser test_9_parser(start = prsd_start, gettoken = prsd_nexttoken, end = prsd_end, lextypes = prsd_lextype);
+CREATE
+
+1: begin;
+BEGIN
+1: create text search configuration test_9_configuration(parser = test_9_parser);
+CREATE
+
+2&: drop text search parser test_9_parser;  <waiting ...>
+
+1: commit;
+COMMIT
+
+2<:  <... completed>
+ERROR:  cannot drop text search parser test_9_parser because other objects depend on it
+DETAIL:  text search configuration test_9_configuration depends on text search parser test_9_parser
+HINT:  Use DROP ... CASCADE to drop the dependent objects too.
+
+1: select count(1) from ts_debug('public.test_9_configuration', 'test');
+ count 
+-------
+ 1     
+(1 row)
+
+drop text search parser test_9_parser cascade;
+DROP
+
+-- Check if dependency is dropped before the creation of the dependent object.
+create text search parser test_9_parser(start = prsd_start, gettoken = prsd_nexttoken, end = prsd_end, lextypes = prsd_lextype);
+CREATE
+
+1: begin;
+BEGIN
+2: begin;
+BEGIN
+2: drop text search parser test_9_parser;
+DROP
+1&: create text search configuration test_9_configuration(parser = test_9_parser);  <waiting ...>
+
+2: commit;
+COMMIT
+1<:  <... completed>
+ERROR:  text search parser ##### was concurrently dropped
+1: end;
+END