ADBDEV-7586 - draft for CI