@RekGRpth 2025-06-30
> Deadlocks are possible. But it can be resolved by the deadlock detection algorithm. So, these 2 transactions trigger ERROR:  deadlock detected with one of the transactions aborted.

Could you add a test for this?