@RekGRpth 2025-06-30
Is it expected that the patch does not cover the case where the type being deleted is used in the function body?
<details><summary>Diff</summary>
<p>

```diff
diff --git a/src/test/isolation2/expected/dependency.out b/src/test/isolation2/expected/dependency.out
index f7a88622630..5d6df295a02 100644
--- a/src/test/isolation2/expected/dependency.out
+++ b/src/test/isolation2/expected/dependency.out
@@ -121,26 +121,28 @@ CREATE
 
 1: begin;
 BEGIN
-1: create function test_3_function(a test_3_type) returns text as $$ select 'Return ' || a; /**/ $$ language sql;
+1: create function test_3_function(a text) returns text as $$ select 'Return ' || a::test_3_type; /**/ $$ language sql;
 CREATE
 
 2&: drop type test_3_type;  <waiting ...>
+FAILED:  Forked command is not blocking; got output: DROP
 
 1: commit;
 COMMIT
 
 2<:  <... completed>
-ERROR:  cannot drop type test_3_type because other objects depend on it
-DETAIL:  function test_3_function(test_3_type) depends on type test_3_type
-HINT:  Use DROP ... CASCADE to drop the dependent objects too.
+FAILED:  Execution failed
 
 1: select test_3_function('one');
- test_3_function 
------------------
- Return one      
-(1 row)
+ERROR:  type "test_3_type" does not exist
+LINE 2:     select 'Return ' || a::test_3_type; /**/
+                                   ^
+QUERY:  
+    select 'Return ' || a::test_3_type; /**/
 
-drop type test_3_type cascade;
+CONTEXT:  SQL function "test_3_function" during inlining
+
+drop function test_3_function(a text) cascade;
 DROP
 
 -- Check if dependency is dropped before the creation of the dependent object.
@@ -152,13 +154,21 @@ BEGIN
 BEGIN
 2: drop type test_3_type;
 DROP
-1&: create function test_3_function(a test_3_type) returns text as $$ select 'Return ' || a; /**/ $$ language sql;  <waiting ...>
+1&: create function test_3_function(a text) returns text as $$ select 'Return ' || a::test_3_type; /**/ $$ language sql;  <waiting ...>
+FAILED:  Forked command is not blocking; got output: CREATE
 
 2: commit;
 COMMIT
 1<:  <... completed>
-ERROR:  cache lookup failed for type XXXXX (lsyscache.c:XXX)
-CONTEXT:  SQL function "test_3_function"
+FAILED:  Execution failed
+1: select test_3_function('one');
+ERROR:  type "test_3_type" does not exist
+LINE 2:     select 'Return ' || a::test_3_type; /**/
+                                   ^
+QUERY:  
+    select 'Return ' || a::test_3_type; /**/
+
+CONTEXT:  SQL function "test_3_function" during inlining
 1: end;
 END
 
diff --git a/src/test/isolation2/sql/dependency.sql b/src/test/isolation2/sql/dependency.sql
index b8bfe65a214..fe3afe39140 100644
--- a/src/test/isolation2/sql/dependency.sql
+++ b/src/test/isolation2/sql/dependency.sql
@@ -84,8 +84,8 @@ $$ language sql;
 create type test_3_type as enum ('one', 'two');
 
 1: begin;
-1: create function test_3_function(a test_3_type) returns text as $$
-    select 'Return ' || a; /**/
+1: create function test_3_function(a text) returns text as $$
+    select 'Return ' || a::test_3_type; /**/
 $$ language sql;
 
 2&: drop type test_3_type;
@@ -96,19 +96,20 @@ $$ language sql;
 
 1: select test_3_function('one');
 
-drop type test_3_type cascade;
+drop function test_3_function(a text) cascade;
 
 -- Check if dependency is dropped before the creation of the dependent object.
 create type test_3_type as enum ('one', 'two');
 1: begin;
 2: begin;
 2: drop type test_3_type;
-1&: create function test_3_function(a test_3_type) returns text as $$
-    select 'Return ' || a; /**/
+1&: create function test_3_function(a text) returns text as $$
+    select 'Return ' || a::test_3_type; /**/
 $$ language sql;
 
 2: commit;
 1<:
+1: select test_3_function('one');
 1: end;
 
 -- Case 4. Function dependency on the language.
```
</p>
</details>