@RekGRpth 2025-06-30
Could you benchmark the performance degradation due to adding new extra locks?