@whitehawk 2025-07-03
@hilltracer 

> Should we add the ticket number into the PR description?

I usually do not add it, as it can be obtained from the branch name. And it is not mandated by the review checklist.

> Maybe add a regression test that checks pg_locks to ensure we don't add extra locks for pinned objects (e.g., int4, text, etc.).

> Maybe also add a test that checks pg_locks to show how the added locks are distributed across our MPP between the coordinator and segments?

I've added check of pg_locks into the first couple of test cases.

---------------

@RekGRpth 

> Could you benchmark the performance degradation due to adding new extra locks?

I've done performance testing with the help of functions below:
```
CREATE OR REPLACE FUNCTION create_many_objects()
RETURNS VOID
AS $$
BEGIN
	for counter in 1..1000 loop
		EXECUTE FORMAT($fmt$CREATE TYPE new_type_%1$s as enum ('one', 'two');$fmt$,
			counter);
		EXECUTE FORMAT($fmt$CREATE TABLE new_table_%1$s (a new_type_%1$s, b new_type_%1$s, c new_type_%1$s) DISTRIBUTED BY (a);$fmt$,
			counter);
	end loop;
END$$
LANGUAGE plpgsql VOLATILE
EXECUTE ON MASTER;

CREATE OR REPLACE FUNCTION remove_many_objects()
RETURNS VOID
AS $$
BEGIN
	-- cleanup
	for counter in 1..1000 loop
		EXECUTE FORMAT($fmt$DROP table new_table_%1$s;$fmt$,
			counter);
		EXECUTE FORMAT($fmt$DROP TYPE new_type_%1$s;$fmt$,
			counter);
	end loop;
END$$
LANGUAGE plpgsql VOLATILE
EXECUTE ON MASTER;
```

commands used for perf measurement:
```
for i in {1..10};
do
	TIME="\t%E sec" time psql postgres -c "select create_many_objects();" > /dev/null
	psql postgres -c "select remove_many_objects();" > /dev/null
done
```

results without the patch:
```
        0:13.64 sec
        0:19.30 sec
        0:22.13 sec
        0:21.60 sec
        0:19.96 sec
        0:20.14 sec
        0:21.98 sec
        0:21.73 sec
        0:22.72 sec
        0:21.62 sec
```
total 204,82 sec.

results with the patch:
```
        0:14.24 sec
        0:21.03 sec
        0:22.06 sec
        0:22.06 sec
        0:20.36 sec
        0:20.16 sec
        0:21.23 sec
        0:21.71 sec
        0:21.60 sec
        0:20.81 sec

```
total 205,26 sec.

So the difference between total numbers looks neglectable. 

> Could you add tests for creating a view that depends on the function being removed, 
> and for changing the column type of a table, and also for the rules that depend on the table being dropped?

Added

> Is it expected that the patch does not cover the case where the object being deleted is used in the function body?

Yes, according to [postgres documentation](https://www.postgresql.org/docs/current/ddl-depend.html):
_For a user-defined function or procedure whose body is defined as a string literal, PostgreSQL tracks dependencies associated with the function's externally-visible properties, such as its argument and result types, but not dependencies that could only be known by examining the function body._

>> Deadlocks are possible. But it can be resolved by the deadlock detection algorithm. So, these 2 transactions trigger ERROR: deadlock detected with one of the transactions aborted.

> Could you add a test for this?

Added.